DIJI_AI/requirements.txt
requests
pyyaml
joblib
pandas
scikit-learn
tensorflow
